let the user cancel the request if the conversation is in a degraded state
