+++ src/script/conversation/MessageRepository.ts
@@ -785,6 +785,40 @@ export class MessageRepository {
   }
 
   /**
+   * Will request user permission before sending a message in case the conversation is in a degraded state
+   *
+   * @param conversation The conversation to send the message in
+   * @returns Resolves to true if the message can be sent, false if the user didn't give their permission
+   */
+  requestUserSendingPermission(conversation: Conversation): Promise<boolean> {
+    if (conversation.verification_state() !== ConversationVerificationState.DEGRADED) {
+      return Promise.resolve(true);
+    }
+    const actionString = t('modalConversationNewDeviceAction');
+    const messageString = t('modalConversationNewDeviceMessage');
+    const titleString = t('modalConversationNewDeviceHeadlineMany'); // TODO get user names
+
+    return new Promise(resolve => {
+      const options: ModalOptions = {
+        close: () => resolve(false),
+        primaryAction: {
+          action: () => {
+            conversation.verification_state(ConversationVerificationState.UNVERIFIED);
+            resolve(true);
+          },
+          text: actionString,
+        },
+        text: {
+          message: messageString,
+          title: titleString,
+        },
+      };
+
+      amplify.publish(WebAppEvents.WARNING.MODAL, ModalsViewModel.TYPE.CONFIRM, options, `degraded-${conversation.id}`);
+    });
+  }
+
+  /**
    * Will send a generic message using @wireapp/code
    *
    * @param payload - the OTR message payload to send
@@ -820,6 +854,8 @@ export class MessageRepository {
       this.cryptography_repository.cryptographyMapper
         .mapGenericMessage(genericMessage, optimisticEvent as EventRecord)
         .then(mappedEvent => this.eventRepository.injectEvent(mappedEvent));
+
+      return this.requestUserSendingPermission(conversation);
     };
 
     const updateOptimisticEvent: MessageSendingCallbacks['onSuccess'] = (genericMessage, sentTime) => {
